R-beta 25w08a
1. The UI of home has been restructured, but some aesthetic fine-tuning is still needed.
2. The communication logic among UI, UIC, and API has been optimized:
Previously, the communication types between UI and UIC were chaotic. For home.py and home_uiC.py, as well as software-related ones, there were two levels of inheritance: class a(Qwidget): and class a_ui(a):. For the login series, it was: class a(object): and class a_ui(QWidget, a):.
After optimization, all the logic has been changed to the login-related category.
3. Fixed some bugs.
4. Further improved the core.py after refactoring.
5. Optimized the error feedback logic during login, introduced a new custom error class inheriting from Exception, avoiding the accidental catching of other errors.
6. Optimized the code logic for some type judgments.